minor BOOST_THROW_EXCEPTION fix.
git-svn-id: http://svn.boost.org/svn/boost/branches/release@58499 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/throw_exception.hpp
@@ -39,7 +39,7 @@
 # define BOOST_THROW_EXCEPTION(x) ::boost::throw_exception(::boost::enable_error_info(x) <<\
     ::boost::throw_function(BOOST_CURRENT_FUNCTION) <<\
     ::boost::throw_file(__FILE__) <<\
-    ::boost::throw_line((int)__LINE__))
+    ::boost::throw_line(__LINE__))
 #else
 # define BOOST_THROW_EXCEPTION(x) ::boost::throw_exception(x)
 #endif
